requirements.txt
# ==============================================================================
# PERSONAL AI EMPLOYEE - PYTHON DEPENDENCIES
# ==============================================================================

# Core
python-dotenv==1.0.0
pydantic==2.5.0

# LLM APIs - Choose one or both
anthropic==0.55.0        # Claude API (Sonnet 4.5)
openai==1.12.0           # OpenAI API (GPT-4, etc.)

# Watchers
watchdog==4.0.0
schedule==1.2.0              # For Monday CEO Briefing automation
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
playwright==1.42.0

# Finance (example: Plaid)
plaid-python==16.0.0

# Orchestration & Async
aiofiles==23.2.1
asyncio==3.4.3

# MCP Protocol
mcp==1.26.0

# Logging & Monitoring
structlog==24.1.0
python-json-logger==2.0.7

# HTTP & API
httpx==0.27.0
requests==2.31.0
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Utilities
pytz==2024.1
dateutil==2.8.2
pyyaml==6.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Security
cryptography==42.0.5

# Optional: Odoo
# odoorpc==0.10.0

# Optional: Slack
# slack-sdk==3.27.0
